+++ plugin/evm/vm.go
@@ -177,6 +177,9 @@ var legacyApiNames = map[string]string{
 // VM implements the snowman.ChainVM interface
 type VM struct {
 	ctx *snow.Context
+	// contextLock is used to coordinate global VM operations.
+	// This can be used safely instead of snow.Context.Lock which is deprecated and should not be used in rpcchainvm.
+	vmLock sync.RWMutex
 	// [cancel] may be nil until [snow.NormalOp] starts
 	cancel context.CancelFunc
 	// *chain.State helps to implement the VM interface by wrapping blocks
@@ -522,7 +525,7 @@ func (vm *VM) Initialize(
 		vm.ctx.ChainID,
 		vm.ctx.WarpSigner,
 		vm,
-		vm.validatorsManager,
+		validators.NewLockedValidatorReader(vm.validatorsManager, &vm.vmLock),
 		vm.warpDB,
 		meteredCache,
 		offchainWarpMessages,
@@ -681,6 +684,8 @@ func (vm *VM) initChainState(lastAcceptedBlock *types.Block) error {
 }
 
 func (vm *VM) SetState(_ context.Context, state snow.State) error {
+	vm.vmLock.Lock()
+	defer vm.vmLock.Unlock()
 	switch state {
 	case snow.StateSyncing:
 		vm.bootstrapped.Set(false)
@@ -721,21 +726,15 @@ func (vm *VM) onNormalOperationsStarted() error {
 	ctx, cancel := context.WithCancel(context.TODO())
 	vm.cancel = cancel
 
-	// sync validators first
-	if err := vm.validatorsManager.Sync(ctx); err != nil {
-		return fmt.Errorf("failed to update validators: %w", err)
+	// Start the validators manager
+	if err := vm.validatorsManager.Initialize(ctx); err != nil {
+		return fmt.Errorf("failed to initialize validators manager: %w", err)
 	}
-	vdrIDs := vm.validatorsManager.GetNodeIDs().List()
-	// Then start tracking with updated validators
-	// StartTracking initializes the uptime tracking with the known validators
-	// and update their uptime to account for the time we were being offline.
-	if err := vm.validatorsManager.StartTracking(vdrIDs); err != nil {
-		return fmt.Errorf("failed to start tracking uptime: %w", err)
-	}
+
 	// dispatch validator set update
 	vm.shutdownWg.Add(1)
 	go func() {
-		vm.validatorsManager.DispatchSync(ctx)
+		vm.validatorsManager.DispatchSync(ctx, &vm.vmLock)
 		vm.shutdownWg.Done()
 	}()
 
@@ -869,6 +868,8 @@ func (vm *VM) setAppRequestHandlers() {
 
 // Shutdown implements the snowman.ChainVM interface
 func (vm *VM) Shutdown(context.Context) error {
+	vm.vmLock.Lock()
+	defer vm.vmLock.Unlock()
 	if vm.ctx == nil {
 		return nil
 	}
@@ -876,12 +877,8 @@ func (vm *VM) Shutdown(context.Context) error {
 		vm.cancel()
 	}
 	if vm.bootstrapped.Get() {
-		vdrIDs := vm.validatorsManager.GetNodeIDs().List()
-		if err := vm.validatorsManager.StopTracking(vdrIDs); err != nil {
-			return fmt.Errorf("failed to stop tracking uptime: %w", err)
-		}
-		if err := vm.validatorsManager.WriteState(); err != nil {
-			return fmt.Errorf("failed to write validator: %w", err)
+		if err := vm.validatorsManager.Shutdown(); err != nil {
+			return fmt.Errorf("failed to shutdown validators manager: %w", err)
 		}
 	}
 	vm.Network.Shutdown()
@@ -1265,6 +1262,9 @@ func attachEthService(handler *rpc.Server, apis []rpc.API, names []string) error
 }
 
 func (vm *VM) Connected(ctx context.Context, nodeID ids.NodeID, version *version.Application) error {
+	vm.vmLock.Lock()
+	defer vm.vmLock.Unlock()
+
 	if err := vm.validatorsManager.Connect(nodeID); err != nil {
 		return fmt.Errorf("uptime manager failed to connect node %s: %w", nodeID, err)
 	}
@@ -1272,6 +1272,9 @@ func (vm *VM) Connected(ctx context.Context, nodeID ids.NodeID, version *version
 }
 
 func (vm *VM) Disconnected(ctx context.Context, nodeID ids.NodeID) error {
+	vm.vmLock.Lock()
+	defer vm.vmLock.Unlock()
+
 	if err := vm.validatorsManager.Disconnect(nodeID); err != nil {
 		return fmt.Errorf("uptime manager failed to disconnect node %s: %w", nodeID, err)
 	}