+++ precompile/contracts/rewardmanager/contract.go
@@ -167,7 +167,7 @@ func GetStoredRewardAddress(stateDB contract.StateDB) (common.Address, bool) {
 	return common.BytesToAddress(val.Bytes()), val == allowFeeRecipientsAddressValue
 }
 
-// StoredRewardAddress stores the given [val] under rewardAddressStorageKey.
+// StoreRewardAddress stores the given [val] under rewardAddressStorageKey.
 func StoreRewardAddress(stateDB contract.StateDB, val common.Address) {
 	stateDB.SetState(ContractAddress, rewardAddressStorageKey, common.BytesToHash(val.Bytes()))
 }