+++ RELEASES.md
@@ -1,13 +1,41 @@
 # Release Notes
 
 ## Pending Release
-* Bump golang version to v1.23.6
-* Bump golangci-lint to v1.63 and add linters
-* Refactored trie_prefetcher.go to be structurally similar to upstream.
-* Remove legacy gossip handler and metrics
+
 * - IMPORTANT: `eth_getProof` calls for historical state will be rejected by default.
   - On archive nodes (`"pruning-enabled": false`): queries for historical proofs for state older than approximately 24 hours preceding the last accepted block will be rejected by default. This can be adjusted with the new option `historical-proof-query-window` which defines the number of blocks before the last accepted block which should be accepted for state proof queries, or set to `0` to accept any block number state query (previous behavior).
   - On `pruning` nodes: queries for proofs past the tip buffer (32 blocks) will be rejected. This is in support of moving to a path based storage scheme, which does not support historical state proofs.
+
+## [v0.7.2](https://github.com/ava-labs/subnet-evm/releases/tag/v0.7.2)
+
+This version is backwards compatible to [v0.7.0](https://github.com/ava-labs/subnet-evm/releases/tag/v0.7.0). It is optional, **but strongly encouraged as it's fixing an important bug in uptime tracking.**
+
+### AvalancheGo Compatibility
+
+The plugin version is unchanged at 39 and is compatible with AvalancheGo version v1.12.2.
+
+### Updates
+
+* Fixed concurrency issue in validators/uptime manager
+* Bump golang version to v1.23.6
+* Bump golangci-lint to v1.63 and add linters
+
+## [v0.7.1](https://github.com/ava-labs/subnet-evm/releases/tag/v0.7.1)
+
+This release focuses on code quality improvements and post-Etna cleanups.
+
+### Compatibility
+
+The plugin version is **updated** to 39 and is compatible with AvalancheGo version v1.12.2.
+
+### Updates
+
+* Moved client type and structs to new `plugin/evm/client` package
+* Fixed statedb improper copy issue
+* Limited the maximum number of query-able rewardPercentile by 100 in `eth_feeHistory` API
+* Refactored `trie_prefetcher.go` to be structurally similar to upstream
+* Removed deprecated legacy gossip handler and metrics
+* Removed unnecessary locks in mempool
 
 ## [v0.7.0](https://github.com/ava-labs/subnet-evm/releases/tag/v0.7.0)